+++ lib/shoulda/matchers/independent/delegate_matcher.rb
@@ -226,7 +226,7 @@ module Shoulda
         end
 
         def subject_has_target_method?
-          subject.respond_to?(target_method)
+          subject.respond_to?(target_method, true)
         end
 
         def ensure_target_method_is_present!